Refactoring & styles fixes
