+++ cvat-ui/src/base.scss
@@ -37,8 +37,6 @@ $objects-bar-icons-color: #242424; // #6e6e6e
 $active-label-background-color: #d8ecff;
 $object-item-border-color: rgba(0, 0, 0, 0.7);
 $slider-color: #1890ff;
-$cansel-button-color1: #b8b8b8;
-$cansel-button-color2: #c7c7c7;
 
 $box-shadow-base: 0 3px 6px -4px rgba(0, 0, 0, 0.12), 0 6px 16px 0 rgba(0, 0, 0, 0.08),
     0 9px 28px 8px rgba(0, 0, 0, 0.05);
