+++ cvat-ui/src/components/create-cloud-storage-page/create-cloud-storage-page.tsx
@@ -6,21 +6,15 @@ import './styles.scss';
 import React from 'react';
 import { Row, Col } from 'antd/lib/grid';
 import Text from 'antd/lib/typography/Text';
-import CreateCloudStorageContent from './create-cloud-storage-content';
 
-interface Props {
-    // error: string;
-    cloudStorageId: number | null;
-}
+import CreateCloudStorageForm from './create-cloud-storage-form';
 
-export default function CreateCloudStoragePageComponent(props: Props): JSX.Element {
-    const { cloudStorageId } = props; // error
-
+export default function CreateCloudStoragePageComponent(): JSX.Element {
     return (
-        <Row justify='center' align='top' className='cvat-create-cloud-storage-form-wrapper'>
+        <Row justify='center' align='top' className='cvat-attach-cloud-storage-form-wrapper'>
             <Col md={20} lg={16} xl={14} xxl={9}>
-                <Text className='cvat-title'>Create a new cloud storage</Text>
-                <CreateCloudStorageContent cloudStorageId={cloudStorageId} />
+                <Text className='cvat-title'>Cloud storage form</Text>
+                <CreateCloudStorageForm />
             </Col>
         </Row>
     );
