+++ cvat-ui/src/utils/enums.tsx
